Make FontHelper-Views testable with robolectric
