+++ qabelbox/src/androidTest/java/de/qabel/qabelbox/ui/helper/UIBoxHelper.java
@@ -61,7 +61,7 @@ public class UIBoxHelper {
         boxManager = helper.getBoxManager();
         identityRepository = helper.getIdentityRepository();
         contactRepository = helper.getContactRepository();
-        preference = helper.getAppPreference();
+        preference = helper.getAppPreferences();
     }
 
     public void uploadFile(Identity identity, String name, byte[] data, String path) throws QblStorageException, IOException {
