+++ qabelbox/src/main/java/de/qabel/qabelbox/views/ButtonFont.java
@@ -6,30 +6,27 @@ import android.widget.Button;
 
 import de.qabel.qabelbox.helper.FontHelper;
 
-/**
- * Created by danny on 12.01.2016.
- */
 public class ButtonFont extends Button {
     public ButtonFont(Context context) {
         super(context);
-        FontHelper.getInstance().setCustomeFonts(this);
+        FontHelper.getInstance().setCustomFonts(this);
     }
 
     public ButtonFont(Context context, AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
-        FontHelper.getInstance().setCustomeFonts(this);
+        FontHelper.getInstance().setCustomFonts(this);
     }
 
 
     public ButtonFont(Context context, AttributeSet attrs) {
         super(context, attrs);
-        FontHelper.getInstance().setCustomeFonts(this);
+        FontHelper.getInstance().setCustomFonts(this);
     }
 
     /*
     @TargetApi(Build.VERSION_CODES.LOLLIPOP)
     public ButtonFont(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
         super(context, attrs, defStyleAttr, defStyleRes);
-        FontHelper.getInstance(context).setCustomeFonts(this);
+        FontHelper.getInstance(context).setCustomFonts(this);
     }*/
 }
