+++ qabelbox/src/main/java/de/qabel/qabelbox/views/EditTextFont.java
@@ -12,23 +12,23 @@ import de.qabel.qabelbox.helper.FontHelper;
 public class EditTextFont extends EditText {
     public EditTextFont(Context context) {
         super(context);
-        FontHelper.getInstance().setCustomeFonts(this);
+        FontHelper.getInstance().setCustomFonts(this);
     }
 
     public EditTextFont(Context context, AttributeSet attrs) {
         super(context, attrs);
-        FontHelper.getInstance().setCustomeFonts(this);
+        FontHelper.getInstance().setCustomFonts(this);
     }
 
     public EditTextFont(Context context, AttributeSet attrs, int defStyleAttr) {
         super(context, attrs, defStyleAttr);
-        FontHelper.getInstance().setCustomeFonts(this);
+        FontHelper.getInstance().setCustomFonts(this);
     }
 
     /*
     @TargetApi(Build.VERSION_CODES.LOLLIPOP)
     public EditTextFont(Context context, AttributeSet attrs, int defStyleAttr, int defStyleRes) {
         super(context, attrs, defStyleAttr, defStyleRes);
-        FontHelper.getInstance(context).setCustomeFonts(this);
+        FontHelper.getInstance(context).setCustomFonts(this);
     }*/
 }
