+++ qabelbox/src/main/java/de/qabel/qabelbox/views/TextViewQabelFont.java
@@ -6,14 +6,11 @@ import android.widget.TextView;
 
 import de.qabel.qabelbox.helper.FontHelper;
 
-/**
- * Created by danny on 12.01.2016.
- */
 public class TextViewQabelFont extends TextView {
 
     public TextViewQabelFont(Context context) {
         super(context);
-        FontHelper.getInstance().setCustomeFonts(this);
+        FontHelper.getInstance().setCustomFonts(this);
     }
 
     public TextViewQabelFont(Context context, AttributeSet attrs) {
